* Switch file pointer offsets to longs instead of ints.

+++ jdem846/src/us/wthr/jdem846/input/gridfloat/GridFloatDataCache.java
@@ -138,7 +138,7 @@ public class GridFloatDataCache extends DataCache
 		Arrays.fill(load_buffer, (byte)0x0);
 		
 
-		int seekStart = start * (Float.SIZE / 8);
+		long seekStart = (long) start * (long) (Float.SIZE / 8);
 		try {
 			inputData.seek(seekStart);
 		} catch (IOException ex) {
